Update Convoy-Site Vertical Offset 📐
The large grid for the convoy site has a "spike" on its bottom to help it stick into
uneven surfaces better. The spike is too long and caused the installation to spawn
several meters above the ground, which makes accessing the door hard.

This change sets an offset of `-3.125` (approx 1.25 blocks) in the `Y` axis to attempt to shove it into the ground more deeply.